Fancy ggplot for SpotRateCurve object

Fancy ggplot for SpotRateCurve object with custom axis, title

ggspotratecurveplot(
  curve,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  curve.name = NULL,
  curve.interpolation = FALSE,
  curve.x.axis = c("dates", "terms"),
  ...
)

Arguments

curve

SpotRateCurve object

title

plot title

subtitle

plot subtitle

caption

plot caption

curve.name

Curve's name, if not provided curve's refdate is used.

curve.interpolation

logical indicating to use daily interpolation instead of curve points. Defaults to FALSE.

curve.x.axis

x axis can be presented with a numeric scale representing business days (terms) or dates (dates). Defaults to dates.

...

additional arguments (not used)

Examples

if (FALSE) { # \dontrun{
terms <- c(1, 11, 26, 27, 28)
rates <- c(0.0719, 0.056, 0.0674, 0.0687, 0.07)
curve <- spotratecurve(rates, terms, "discrete", "actual/365", "actual")
ggspotratecurveplot(curve)
} # }