SpotRate class

The SpotRate class abstracts the interst rate and has methods to handle many calculations on it.

Details

The SpotRate class fully specifies spot rates. It has:

the spot rate values which are numeric values representing the rate.

the compounding regime that specifies how to compound the spot rate. This is a Compounding object.

the daycount rule to compute the compounding periods right adjusted to the spot rate frequency.

the calendar according to which the number of days are counted.

The SpotRate class is a numeric, that represents the interest rate and that has the slots: compounding, daycount and calendar.

For example, an annual simple interest rate of 6%, that compounds in calendar days, is defined as follows:

sr_simple <- spotrate(0.06, "simple", "actual/360", "actual")
sr_simple

## [1] "0.06 simple actual/360 actual"

actual/360 is the daycount rule and actual is the calendar.

Differently, an annual compound interest rate of 10%, that compounds in business days according to calendar Brazil/ANBIMA is

sr_disc <- spotrate(0.1, "discrete", "business/252", "Brazil/ANBIMA")
sr_disc

## [1] "0.1 discrete business/252 Brazil/ANBIMA"

The calendar slot is a bizdays calendar.

An $100,000 investment in an instrument that pays that interst rate for 5 years has the future value.

100000 * compound(sr_disc, term(5, "years"))

## [1] 161051

for the simple interest rate we have

100000 * compound(sr_simple, term(5, "years"))

## [1] 130000

SpotRate objects can be created with vectors

sr_vec <- spotrate(abs(rnorm(10)), "discrete", "business/252", "Brazil/ANBIMA")
sr_vec

##  [1] "2.4074949 discrete business/252 Brazil/ANBIMA"
##  [2] "0.1536445 discrete business/252 Brazil/ANBIMA"
##  [3] "0.7272605 discrete business/252 Brazil/ANBIMA"
##  [4] "0.1665265 discrete business/252 Brazil/ANBIMA"
##  [5] "0.1888282 discrete business/252 Brazil/ANBIMA"
##  [6] "0.8619113 discrete business/252 Brazil/ANBIMA"
##  [7] "0.3456210 discrete business/252 Brazil/ANBIMA"
##  [8] "1.1039747 discrete business/252 Brazil/ANBIMA"
##  [9] "0.4384621 discrete business/252 Brazil/ANBIMA"
## [10] "1.7957924 discrete business/252 Brazil/ANBIMA"

and can be put into a data.frame

data.frame(spot_rate = sr_vec)

##                                        spot_rate
## 1  2.4074949 discrete business/252 Brazil/ANBIMA
## 2  0.1536445 discrete business/252 Brazil/ANBIMA
## 3  0.7272605 discrete business/252 Brazil/ANBIMA
## 4  0.1665265 discrete business/252 Brazil/ANBIMA
## 5  0.1888282 discrete business/252 Brazil/ANBIMA
## 6  0.8619113 discrete business/252 Brazil/ANBIMA
## 7  0.3456210 discrete business/252 Brazil/ANBIMA
## 8  1.1039747 discrete business/252 Brazil/ANBIMA
## 9  0.4384621 discrete business/252 Brazil/ANBIMA
## 10 1.7957924 discrete business/252 Brazil/ANBIMA

once in a data.frame, dplyr verbs can be used to manipulate it.

require(dplyr, warn.conflicts = FALSE)

## Loading required package: dplyr

data.frame(spot_rate = sr_vec) |>
   mutate(comp = compound(spot_rate, term(5, "months")))

##                                        spot_rate     comp
## 1  2.4074949 discrete business/252 Brazil/ANBIMA 1.666664
## 2  0.1536445 discrete business/252 Brazil/ANBIMA 1.061362
## 3  0.7272605 discrete business/252 Brazil/ANBIMA 1.255738
## 4  0.1665265 discrete business/252 Brazil/ANBIMA 1.066284
## 5  0.1888282 discrete business/252 Brazil/ANBIMA 1.074731
## 6  0.8619113 discrete business/252 Brazil/ANBIMA 1.295636
## 7  0.3456210 discrete business/252 Brazil/ANBIMA 1.131665
## 8  1.1039747 discrete business/252 Brazil/ANBIMA 1.363328
## 9  0.4384621 discrete business/252 Brazil/ANBIMA 1.163566
## 10 1.7957924 discrete business/252 Brazil/ANBIMA 1.534771

SpotRate is numeric, so it executes arithmetic and comparison operations with numeric objects.

data.frame(spot_rate = sr_vec) |>
   mutate(
     new_spot_rate = spot_rate + 0.02,
     check_gt_1pp = spot_rate > 0.01,
     check_gt_nsr = spot_rate > new_spot_rate
   )

##                                        spot_rate
## 1  2.4074949 discrete business/252 Brazil/ANBIMA
## 2  0.1536445 discrete business/252 Brazil/ANBIMA
## 3  0.7272605 discrete business/252 Brazil/ANBIMA
## 4  0.1665265 discrete business/252 Brazil/ANBIMA
## 5  0.1888282 discrete business/252 Brazil/ANBIMA
## 6  0.8619113 discrete business/252 Brazil/ANBIMA
## 7  0.3456210 discrete business/252 Brazil/ANBIMA
## 8  1.1039747 discrete business/252 Brazil/ANBIMA
## 9  0.4384621 discrete business/252 Brazil/ANBIMA
## 10 1.7957924 discrete business/252 Brazil/ANBIMA
##                                    new_spot_rate check_gt_1pp check_gt_nsr
## 1  2.4274949 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 2  0.1736445 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 3  0.7472605 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 4  0.1865265 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 5  0.2088282 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 6  0.8819113 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 7  0.3656210 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 8  1.1239747 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 9  0.4584621 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 10 1.8157924 discrete business/252 Brazil/ANBIMA         TRUE        FALSE

Note

The SpotRate objects are annual rates.