SpotRate class

The SpotRate class abstracts the interst rate and has methods to handle many calculations on it.

Details

The SpotRate class fully specifies spot rates. It has:

the spot rate values which are numeric values representing the rate.

the compounding regime that specifies how to compound the spot rate. This is a Compounding object.

the daycount rule to compute the compounding periods right adjusted to the spot rate frequency.

the calendar according to which the number of days are counted.

The SpotRate class is a numeric, that represents the interest rate and that has the slots: compounding, daycount and calendar.

For example, an annual simple interest rate of 6%, that compounds in calendar days, is defined as follows:

sr_simple <- spotrate(0.06, "simple", "actual/360", "actual")
sr_simple

## [1] "0.06 simple actual/360 actual"

actual/360 is the daycount rule and actual is the calendar.

Differently, an annual compound interest rate of 10%, that compounds in business days according to calendar Brazil/ANBIMA is

sr_disc <- spotrate(0.1, "discrete", "business/252", "Brazil/ANBIMA")
sr_disc

## [1] "0.1 discrete business/252 Brazil/ANBIMA"

The calendar slot is a bizdays calendar.

An $100,000 investment in an instrument that pays that interst rate for 5 years has the future value.

100000 * compound(sr_disc, term(5, "years"))

## [1] 161051

for the simple interest rate we have

100000 * compound(sr_simple, term(5, "years"))

## [1] 130000

SpotRate objects can be created with vectors

sr_vec <- spotrate(abs(rnorm(10)), "discrete", "business/252", "Brazil/ANBIMA")
sr_vec

##  [1] "0.71537412 discrete business/252 Brazil/ANBIMA"
##  [2] "0.23891015 discrete business/252 Brazil/ANBIMA"
##  [3] "0.41318226 discrete business/252 Brazil/ANBIMA"
##  [4] "0.73443121 discrete business/252 Brazil/ANBIMA"
##  [5] "0.06175294 discrete business/252 Brazil/ANBIMA"
##  [6] "0.66272197 discrete business/252 Brazil/ANBIMA"
##  [7] "0.49816241 discrete business/252 Brazil/ANBIMA"
##  [8] "0.41765945 discrete business/252 Brazil/ANBIMA"
##  [9] "0.42891369 discrete business/252 Brazil/ANBIMA"
## [10] "0.09361829 discrete business/252 Brazil/ANBIMA"

and can be put into a data.frame

data.frame(spot_rate = sr_vec)

##                                         spot_rate
## 1  0.71537412 discrete business/252 Brazil/ANBIMA
## 2  0.23891015 discrete business/252 Brazil/ANBIMA
## 3  0.41318226 discrete business/252 Brazil/ANBIMA
## 4  0.73443121 discrete business/252 Brazil/ANBIMA
## 5  0.06175294 discrete business/252 Brazil/ANBIMA
## 6  0.66272197 discrete business/252 Brazil/ANBIMA
## 7  0.49816241 discrete business/252 Brazil/ANBIMA
## 8  0.41765945 discrete business/252 Brazil/ANBIMA
## 9  0.42891369 discrete business/252 Brazil/ANBIMA
## 10 0.09361829 discrete business/252 Brazil/ANBIMA

once in a data.frame, dplyr verbs can be used to manipulate it.

require(dplyr, warn.conflicts = FALSE)

## Loading required package: dplyr

data.frame(spot_rate = sr_vec) |>
   mutate(comp = compound(spot_rate, term(5, "months")))

##                                         spot_rate     comp
## 1  0.71537412 discrete business/252 Brazil/ANBIMA 1.252130
## 2  0.23891015 discrete business/252 Brazil/ANBIMA 1.093369
## 3  0.41318226 discrete business/252 Brazil/ANBIMA 1.155002
## 4  0.73443121 discrete business/252 Brazil/ANBIMA 1.257908
## 5  0.06175294 discrete business/252 Brazil/ANBIMA 1.025281
## 6  0.66272197 discrete business/252 Brazil/ANBIMA 1.235971
## 7  0.49816241 discrete business/252 Brazil/ANBIMA 1.183449
## 8  0.41765945 discrete business/252 Brazil/ANBIMA 1.156525
## 9  0.42891369 discrete business/252 Brazil/ANBIMA 1.160342
## 10 0.09361829 discrete business/252 Brazil/ANBIMA 1.037992

SpotRate is numeric, so it executes arithmetic and comparison operations with numeric objects.

data.frame(spot_rate = sr_vec) |>
   mutate(
     new_spot_rate = spot_rate + 0.02,
     check_gt_1pp = spot_rate > 0.01,
     check_gt_nsr = spot_rate > new_spot_rate
   )

##                                         spot_rate
## 1  0.71537412 discrete business/252 Brazil/ANBIMA
## 2  0.23891015 discrete business/252 Brazil/ANBIMA
## 3  0.41318226 discrete business/252 Brazil/ANBIMA
## 4  0.73443121 discrete business/252 Brazil/ANBIMA
## 5  0.06175294 discrete business/252 Brazil/ANBIMA
## 6  0.66272197 discrete business/252 Brazil/ANBIMA
## 7  0.49816241 discrete business/252 Brazil/ANBIMA
## 8  0.41765945 discrete business/252 Brazil/ANBIMA
## 9  0.42891369 discrete business/252 Brazil/ANBIMA
## 10 0.09361829 discrete business/252 Brazil/ANBIMA
##                                     new_spot_rate check_gt_1pp check_gt_nsr
## 1  0.73537412 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 2  0.25891015 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 3  0.43318226 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 4  0.75443121 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 5  0.08175294 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 6  0.68272197 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 7  0.51816241 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 8  0.43765945 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 9  0.44891369 discrete business/252 Brazil/ANBIMA         TRUE        FALSE
## 10 0.11361829 discrete business/252 Brazil/ANBIMA         TRUE        FALSE

SpotRate vectors also are created with the concatenation function c.

c(sr_disc, 0.1, 0.13, 0.14, 0.,15)

## [1] " 0.10 discrete business/252 Brazil/ANBIMA"
## [2] " 0.10 discrete business/252 Brazil/ANBIMA"
## [3] " 0.13 discrete business/252 Brazil/ANBIMA"
## [4] " 0.14 discrete business/252 Brazil/ANBIMA"
## [5] " 0.00 discrete business/252 Brazil/ANBIMA"
## [6] "15.00 discrete business/252 Brazil/ANBIMA"

Furtherly, all indexing operations of numeric objects are supported by SpotRate objects.

Invalid Operations

Operations involving SpotRate objects with different compounding, daycount or calendar, raise errors.

This happens with the following operations:

Compare: >, <, <=, >=, ==, !=

Arithmetic: +, -, *, /

Concatenation: c

try(sr_simple + sr_disc)

## Error in stop_if_spotrate_slots_differ(e1, e2, "SpotRate objects have different slots") : 
##   SpotRate objects have different slots

try(sr_simple > sr_disc)

## [1] FALSE

try(c(sr_simple, sr_disc))

## Error in stop_if_spotrate_slots_differ(x, values_, "SpotRate objects have different slots") : 
##   SpotRate objects have different slots

Note

The SpotRate objects are annual rates.