SpotRate class

The SpotRate class abstracts the interst rate and has methods to handle many calculations on it.

Details

The SpotRate class fully specifies spot rates. It has:

the spot rate values which are numeric values representing the rate.

the compounding regime that specifies how to compound the spot rate. This is a Compounding object.

the daycount rule to compute the compounding periods right adjusted to the spot rate frequency (which is annual).

the calendar according to which the number of days are counted.

The SpotRate class is a numeric, that represents the interest rate and that has the slots: compounding, daycount and calendar.

The calendar slot is a bizdays calendar.

Note

The SpotRate objects are annual rates.